ADD montage.py, consolidate Montages into one file
