+++ mne/channels/__init__.py
@@ -4,6 +4,7 @@ setting of sensors locations used for processing and plotting.
 """
 
 from .layout import (Layout, make_eeg_layout, make_grid_layout, read_layout,
-                     find_layout, read_montage)
+                     find_layout)
+from .montage import read_montage, read_dig_montage, Montage, DigMontage
 
 from .channels import (equalize_channels, rename_channels, read_ch_connectivity)
