+++ mne/io/__init__.py
@@ -6,8 +6,7 @@
 # License: BSD (3-clause)
 
 from .open import fiff_open, show_fiff, _fiff_get_fid
-from .meas_info import (read_fiducials, write_fiducials, read_info, write_info,
-                        read_dig_montage)
+from .meas_info import read_fiducials, write_fiducials, read_info, write_info
 
 from .proj import make_eeg_average_ref_proj
 from . import array
@@ -35,4 +34,3 @@ from .fiff import RawFIF as Raw
 from .base import concatenate_raws
 from .chpi import  get_chpi_positions
 from .reference import set_eeg_reference, set_bipolar_reference, add_reference_channels
-
