+++ mne/io/meas_info.py
@@ -1229,182 +1229,3 @@ def _empty_info():
                           'to': FIFF.FIFFV_COORD_HEAD, 'trans': np.eye(4)}
     assert set(info.keys()) == set(RAW_INFO_FIELDS)
     return info
-
-
-class DigMontage(object):
-    """Montage for Digitized data
-
-    Montages are typically loaded from a file using read_montage. Only use this
-    class directly if you're constructing a new montage.
-
-    Parameters
-    ----------
-    hsp : array, shape (n_points, 3)
-        The positions of the channels in 3d.
-    hpi : array, shape (n_hpi, 3)
-        The positions of the head-position indicator coils in 3d.
-        These points are in the MEG device space.
-    elp : array, shape (n_hpi, 3)
-        The positions of the head-position indicator coils in 3d.
-        This is typically in the acquisition digitizer space.
-    point_names : list, shape (n_elp)
-        The names of the digitized points for hpi and elp.
-    nasion : array, shape (1, 3)
-        The position of the nasion fidicual point in the RAS head space.
-    lpa : array, shape (1, 3)
-        The position of the left periauricular fidicual point in
-        the RAS head space.
-    rpa : array, shape (1, 3)
-        The position of the right periauricular fidicual point in
-        the RAS head space.
-    dev_head_t : array, shape (4, 4)
-        A Dev-to-Head transformation matrix.
-    """
-    def __init__(self, hsp, hpi, elp, point_names,
-                 nasion=None, lpa=None, rpa=None, dev_head_t=None):
-        self.hsp = hsp
-        self.hpi = hpi
-        self.elp = elp
-        self.point_names = point_names
-
-        self.nasion = nasion
-        self.lpa = lpa
-        self.rpa = rpa
-        if dev_head_t is None:
-            self.dev_head_t = np.identity(4)
-        else:
-            self.dev_head_t = dev_head_t
-
-    def __repr__(self):
-        s = '<DigMontage | %d Dig Points, %d HPI points: %s ...>'
-        s %= (len(self.hsp), len(self.point_names),
-              ', '.join(self.point_names[:3]))
-        return s
-
-    def plot(self, scale_factor=1.5, show_names=False):
-        """Plot EEG sensor montage
-
-        Parameters
-        ----------
-        scale_factor : float
-            Determines the size of the points. Defaults to 1.5
-        show_names : bool
-            Whether to show the channel names. Defaults to False
-
-        Returns
-        -------
-        fig : Instance of matplotlib.figure.Figure
-            The figure object.
-        """
-        from ..viz import plot_montage
-        return plot_montage(self, scale_factor=scale_factor,
-                            show_names=show_names)
-
-
-def read_dig_montage(hsp=None, hpi=None, elp=None, point_names=None,
-                     unit='mm', transform=True, dev_head_t=False):
-    """Read montage from a file
-
-    Note: Digitized points will be transformed to head-based coordinate system.
-
-    Parameters
-    ----------
-    hsp : None | str | array, shape (n_points, 3)
-        If str, this corresponds to the filename of the headshape points.
-        This is typically used with the Polhemus FastSCAN system.
-        If numpy.array, this corresponds to an array of positions of the
-        channels in 3d.
-    hpi : None | str | array, shape (n_hpi, 3)
-        If str, this corresponds to the filename of hpi points. If numpy.array,
-        this corresponds to an array hpi points. These points are in
-        device space.
-    elp : None | str | array, shape (n_fids + n_hpi, 3)
-        If str, this corresponds to the filename of hpi points.
-        This is typically used with the Polhemus FastSCAN system.
-        If numpy.array, this corresponds to an array hpi points. These points
-        are in head space. Fiducials should be listed first, then the points
-        corresponding to the hpi.
-    point_names : None | list
-        If list, this corresponds to a list of point names. This must be
-        specified if elp is defined.
-    unit : 'm' | 'cm' | 'mm'
-        Unit of the input file. If not 'm', coordinates will be rescaled
-        to 'm'. Default is 'mm'. This is applied only for hsp and elp files.
-    transform : bool
-        If True, points will be transformed to Neuromag space.
-        The fidicuals, 'nasion', 'lpa', 'rpa' must be specified in
-        the montage file. Useful for points captured using Polhemus FastSCAN.
-        Default is True.
-    dev_head_t : bool
-        If True, a Dev-to-Head transformation matrix will be added to the
-        montage. To get a proper `dev_head_t`, the hpi and the elp points
-        must be in the same order. If False, an identity matrix will be added
-        to the montage. Default is False.
-
-
-    Returns
-    -------
-    montage : instance of DigMontage
-        The digitizer montage.
-    """
-    if isinstance(hsp, string_types):
-        hsp = _read_dig_points(hsp)
-    if hsp is not None:
-        if unit == 'mm':
-            hsp *= 1e-3
-        if unit == 'cm':
-            hsp *= 1e-2
-    if isinstance(hpi, string_types):
-        ext = op.splitext(hpi)[-1]
-        if ext == '.txt':
-            hpi = _read_dig_points(hpi)
-        elif ext in ('.sqd', '.mrk'):
-            from .kit import read_mrk
-            hpi = read_mrk(hpi)
-        else:
-            raise TypeError('HPI file is not supported.')
-    if isinstance(elp, string_types):
-        elp = _read_dig_points(elp)
-    if elp is not None:
-        if len(elp) != len(point_names):
-            raise ValueError("The elp file contains %i points, but %i names "
-                             "were specified." % (len(elp), len(point_names)))
-        if unit == 'mm':
-            elp *= 1e-3
-        elif unit == 'cm':
-            elp *= 1e-2
-
-    if transform:
-        if elp is None:
-            raise ValueError("ELP points are not specified. Points are needed "
-                             "for transformation.")
-        names_lower = [name.lower() for name in point_names]
-
-        # check that all needed points are present
-        missing = tuple(name for name in ('nasion', 'lpa', 'rpa')
-                        if name not in names_lower)
-        if missing:
-            raise ValueError("The points %s are missing, but are needed "
-                             "to transform the points to the MNE coordinate "
-                             "system. Either add the points, or read the "
-                             "montage with transform=False." % str(missing))
-
-        nasion = elp[names_lower.index('nasion')]
-        lpa = elp[names_lower.index('lpa')]
-        rpa = elp[names_lower.index('rpa')]
-        neuromag_trans = get_ras_to_neuromag_trans(nasion, lpa, rpa)
-
-        fids = np.array([nasion, lpa, rpa])
-        fids = apply_trans(neuromag_trans, fids)
-        elp = apply_trans(neuromag_trans, elp)
-        hsp = apply_trans(neuromag_trans, hsp)
-    else:
-        fids = [None] * 3
-    if dev_head_t:
-        from ..coreg import fit_matched_points
-        trans = fit_matched_points(tgt_pts=elp[3:], src_pts=hpi, out='trans')
-    else:
-        trans = np.identity(4)
-
-    return DigMontage(hsp, hpi, elp, point_names, fids[0], fids[1], fids[2],
-                      trans)
